refactor: Phase 3 split types.ts, unify Logger with Strategy pattern, consolidate Error hierarchy
- Split types.ts (646 lines, 51 types) into 5 domain files with export * backward compat
- Create shared/logging/ with LogWriter interface, JsonlWriter, TextWriter, FilePerEntryWriter
- Migrate 3 logger systems to use shared LogWriter strategy pattern
- Create shared/errors/ with ClimptError base class and domain-specific error modules
- Consolidate error re-exports in runner/errors.ts for backward compatibility
