+++ agents/src_common/types/mod.ts
@@ -1,0 +1,10 @@
+/**
+ * Re-export all type definitions for backward compatibility.
+ * Import from individual files for domain-specific types.
+ */
+
+export * from "./agent-definition.ts";
+export * from "./completion.ts";
+export * from "./step-flow.ts";
+export * from "./validation.ts";
+export * from "./runtime.ts";
